Filesystem: add generic abstraction layer
This change separates the generic filesystem code from the TFS
code, and moves the generic code to an abstraction layer that sits
between the TFS code and the rest of the kernel.
Ths src/tfs directory is being replaced by src/fs, which contains
the generic filesystem code as well as different filesystem
implementations (currently only TFS).
In order to allow future filesystem implementations where fsfile
structures may be allocated and deallocated on-demand based on the
user program or kernel accessing the files (as opposed to being
allocated when a file is created and deallocated when a file is
deleted), any function that returns a reference to an fsfile
increments the refcount of the fsfile, and the caller has to
decrement the refcount when the fsfile is no longer used.
Use of TFS-specific inerfaces outside the TFS code is being
replaced by calls to generic filesystem functions; the filesystem
and fsfile structures no longer have TFS-specific fields, but have
a set of pointers to callback functions, for which each filesystem
provides its implementation.
Filesystem callbacks are called with the filesystem lock held.
