+++ platform/riscv-virt/Makefile
@@ -47,8 +47,9 @@ SRCS-kernel.elf= \
 	$(SRCDIR)/net/net.c \
 	$(SRCDIR)/net/netsyscall.c \
 	$(RUNTIME) \
-	$(SRCDIR)/tfs/tfs.c \
-	$(SRCDIR)/tfs/tlog.c \
+	$(SRCDIR)/fs/fs.c \
+	$(SRCDIR)/fs/tfs.c \
+	$(SRCDIR)/fs/tlog.c \
 	$(SRCDIR)/unix/aio.c \
 	$(SRCDIR)/unix/blockq.c \
 	$(SRCDIR)/unix/coredump.c \
@@ -131,7 +132,7 @@ INCLUDES=\
 	-I$(SRCDIR)/kernel \
 	-I$(SRCDIR)/net \
 	-I$(SRCDIR)/runtime \
-	-I$(SRCDIR)/tfs \
+	-I$(SRCDIR)/fs \
 	-I$(SRCDIR)/unix \
 	-I$(SRCDIR)/unix_process \
 	-I$(SRCDIR)/xen/public \
