+++ src/fs/fs.h
@@ -6,32 +6,11 @@ typedef closure_type(filesystem_complete, void, filesystem, status);
 
 typedef struct fsfile *fsfile;
 
-filesystem fsfile_get_fs(fsfile f);
-pagecache_volume filesystem_get_pagecache_volume(filesystem fs);
-
 u64 fsfile_get_length(fsfile f);
 void fsfile_set_length(fsfile f, u64 length);
 sg_io fsfile_get_reader(fsfile f);
 sg_io fsfile_get_writer(fsfile f);
 pagecache_node fsfile_get_cachenode(fsfile f);
-
-extern io_status_handler ignore_io_status;
-
-#define MIN_EXTENT_SIZE PAGESIZE
-#define MIN_EXTENT_ALLOC_SIZE   (1 * MB)
-
-boolean filesystem_probe(u8 *first_sector, u8 *uuid, char *label);
-const char *filesystem_get_label(filesystem fs);
-void filesystem_get_uuid(filesystem fs, u8 *uuid);
-
-void create_filesystem(heap h,
-                       u64 blocksize,
-                       u64 size,
-                       storage_req_handler req_handler,
-                       boolean ro,
-                       const char *label,
-                       filesystem_complete complete);
-void destroy_filesystem(filesystem fs);
 
 // turn these into method gets rather than call
 void filesystem_read_sg(fsfile f, sg_list sg, range q, status_handler completion);
@@ -60,16 +39,12 @@ void filesystem_set_mtime(filesystem fs, tuple t, timestamp tim);
 u64 filesystem_get_rdev(filesystem fs, tuple t);
 void filesystem_set_rdev(filesystem fs, tuple t, u64 rdev);
 
-u64 fsfile_get_length(fsfile f);
-void fsfile_set_length(fsfile f, u64);
-u64 fsfile_get_blocks(fsfile f);    /* returns the number of allocated blocks */
-fsfile fsfile_from_node(filesystem fs, tuple n);
-fsfile file_lookup(filesystem fs, vector v);
 void filesystem_read_entire(filesystem fs, tuple t, heap bufheap, buffer_handler c, status_handler s);
-fsfile allocate_fsfile(filesystem fs, tuple md);
 void fsfile_reserve(fsfile f);
 void fsfile_release(fsfile f);
 void fsfile_flush(fsfile f, boolean datasync, status_handler completion);
+
+#define fsfile_get_blocks(f)    (f)->get_blocks(f)  /* returns the number of allocated blocks */
 
 typedef enum {
     FS_STATUS_OK = 0,
@@ -91,9 +66,6 @@ typedef enum {
 
 const char *string_from_fs_status(fs_status s);
 
-fs_status filesystem_write_tuple(filesystem fs, tuple t);
-fs_status filesystem_write_eav(filesystem fs, tuple t, symbol a, value v);
-
 typedef closure_type(fs_status_handler, void, fsfile, fs_status);
 
 void filesystem_alloc(fsfile f, long offset, long len,
@@ -101,14 +73,97 @@ void filesystem_alloc(fsfile f, long offset, long len,
 void filesystem_dealloc(fsfile f, long offset, long len,
         fs_status_handler completion);
 fs_status filesystem_truncate(filesystem fs, fsfile f, u64 len);
+fs_status filesystem_truncate_locked(filesystem fs, fsfile f, u64 len);
 
-fs_status do_mkentry(filesystem fs, tuple parent, const char *name, tuple entry,
-        boolean persistent);
+fs_status fsfile_init(filesystem fs, fsfile f, tuple md, sg_io fs_read, sg_io fs_write,
+                      pagecache_node_reserve fs_reserve, thunk fs_free);
 
-fs_status filesystem_mkentry(filesystem fs, tuple cwd, const char *fp, tuple entry,
-    boolean persistent, boolean recursive);
-fs_status filesystem_mkdirpath(filesystem fs, tuple cwd, const char *fp,
-        boolean persistent);
+declare_closure_struct(1, 0, void, fs_sync,
+                       filesystem, fs);
+declare_closure_struct(1, 1, void, fs_free,
+                       filesystem, fs,
+                       status, s);
+struct filesystem {
+    u64 size;
+    heap h;
+    int blocksize_order;
+    boolean ro; /* true for read-only filesystem */
+    pagecache_volume pv;
+    tuple (*lookup)(filesystem fs, tuple parent, string name);
+    fs_status (*create)(filesystem fs, tuple parent, string name, tuple md, fsfile *f);
+    fs_status (*unlink)(filesystem fs, tuple parent, string name, tuple md, boolean *destruct_md);
+    fs_status (*rename)(filesystem fs, tuple old_parent, string old_name, tuple old_md,
+                        tuple new_parent, string new_name, tuple new_md, boolean exchange,
+                        boolean *destruct_md);
+    fs_status (*truncate)(filesystem fs, fsfile f, u64 len);
+    fs_status (*get_fsfile)(filesystem fs, tuple md, fsfile *f);
+    inode (*get_inode)(filesystem fs, tuple md);
+    tuple (*get_meta)(filesystem fs, inode n);
+    u64 (*get_freeblocks)(filesystem fs);
+    status_handler (*get_sync_handler)(filesystem fs, fsfile fsf, boolean datasync,
+                                       status_handler completion);
+    void (*destroy_fs)(filesystem fs);
+    tuple root;
+#ifdef KERNEL
+    struct spinlock lock;
+#endif
+    struct refcount refcount;
+    closure_struct(fs_sync, sync);
+    thunk sync_complete;
+    closure_struct(fs_free, free);
+};
+
+declare_closure_struct(1, 1, void, fsf_sync_complete,
+                       fsfile, f,
+                       status, s);
+struct fsfile {
+    filesystem fs;
+    pagecache_node cache_node;
+    u64 length;
+    tuple md;
+    sg_io read;
+    sg_io write;
+    s64 (*get_blocks)(fsfile f);
+    struct refcount refcount;
+    closure_struct(fsf_sync_complete, sync_complete);
+    u8 status;
+};
+
+/* fsfile status flags */
+#define FSF_DIRTY_DATASYNC  (1 << 0)    /* metadata needed for retrieving file data */
+#define FSF_DIRTY_OTHER     (1 << 1)    /* any other metadata */
+#define FSF_DIRTY           (FSF_DIRTY_DATASYNC | FSF_DIRTY_OTHER)
+
+status filesystem_init(filesystem fs, heap h, u64 size, u64 blocksize, boolean ro);
+
+static inline u64 bytes_from_sectors(filesystem fs, u64 sectors)
+{
+    return sectors << fs->blocksize_order;
+}
+
+static inline u64 sector_from_offset(filesystem fs, bytes b)
+{
+    return b >> fs->blocksize_order;
+}
+
+#ifdef KERNEL
+
+#define filesystem_lock_init(fs)    spin_lock_init(&(fs)->lock)
+#define filesystem_lock(fs)         spin_lock(&(fs)->lock)
+#define filesystem_unlock(fs)       spin_unlock(&(fs)->lock)
+
+#else
+
+#define filesystem_lock_init(fs)
+#define filesystem_lock(fs)         ((void)fs)
+#define filesystem_unlock(fs)       ((void)fs)
+
+#endif
+
+tuple fs_new_entry(filesystem fs);
+
+boolean file_tuple_is_ancestor(tuple t1, tuple t2, tuple p2);
+
 fs_status filesystem_mkdir(filesystem fs, inode cwd, const char *path);
 fs_status filesystem_get_node(filesystem *fs, inode cwd, const char *path, boolean nofollow,
                               boolean create, boolean exclusive, boolean truncate, tuple *n,
@@ -146,10 +201,6 @@ extern const char *gitversion;
 #define NAME_MAX 255
 #define PATH_MAX 4096
 
-static inline inode inode_from_tuple(tuple t) {
-    return u64_from_pointer(t);
-}
-
 static inline buffer linktarget(tuple x)
 {
     return get_string(x, sym(linktarget));
@@ -168,6 +219,16 @@ static inline boolean is_symlink(tuple n)
 static inline boolean is_socket(tuple n)
 {
     return get(n, sym(socket)) ? true : false;
+}
+
+static inline boolean is_special(tuple n)
+{
+    return get(n, sym(special)) ? true : false;
+}
+
+static inline boolean is_regular(tuple n)
+{
+    return (!is_dir(n) && !is_symlink(n) && !is_socket(n) && !is_special(n));
 }
 
 static inline char *path_find_last_delim(const char *path, unsigned int len)
@@ -206,7 +267,7 @@ int file_get_path(filesystem fs, inode ino, char *buf, u64 len);
 
 #ifdef KERNEL
 
-/* Functions called by TFS code to notify filesystem operations */
+/* Functions called by filesystem code to notify filesystem operations */
 void fs_notify_create(tuple t, tuple parent, symbol name);
 void fs_notify_move(tuple t, tuple old_parent, symbol old_name, tuple new_parent, symbol new_name);
 void fs_notify_delete(tuple t, tuple parent, symbol name);
@@ -222,3 +283,5 @@ void fs_notify_release(tuple t, boolean unmounted);
 #define fs_notify_release(t, u)             (void)(t)
 
 #endif
+
+boolean fs_file_is_busy(filesystem fs, tuple md);
