+++ src/kernel/init.c
@@ -175,7 +175,7 @@ closure_function(2, 2, void, fsstarted,
     closure_finish();
     symbol booted = sym(booted);
     if (!get(root, booted))
-        filesystem_write_eav(fs, root, booted, null_value);
+        filesystem_write_eav((tfs)fs, root, booted, null_value);
     config_console(root);
 }
