+++ src/kernel/tracelog.c
@@ -640,7 +640,7 @@ static void init_tracelog_file_writer(value v)
     }
     filesystem fs = get_root_fs();
     tuple root = filesystem_getroot(fs);
-    fs_status s = filesystem_get_node(&fs, inode_from_tuple(root),
+    fs_status s = filesystem_get_node(&fs, fs->get_inode(fs, root),
                                       buffer_to_cstring((buffer)v),
                                       true, true, false, false, &file, &fsf);
     if (s != FS_STATUS_OK) {
