+++ tools/Makefile
@@ -8,16 +8,18 @@ SRCS-dump= \
 	$(CURDIR)/dump.c \
 	$(SRCDIR)/kernel/pagecache.c \
 	$(RUNTIME) \
-	$(SRCDIR)/tfs/tfs.c \
-	$(SRCDIR)/tfs/tlog.c \
+	$(SRCDIR)/fs/fs.c \
+	$(SRCDIR)/fs/tfs.c \
+	$(SRCDIR)/fs/tlog.c \
 	$(SRCDIR)/unix_process/unix_process_runtime.c
 
 SRCS-tfs-fuse= \
 	$(CURDIR)/tfs-fuse.c \
 	$(SRCDIR)/kernel/pagecache.c \
 	$(RUNTIME) \
-	$(SRCDIR)/tfs/tfs.c \
-	$(SRCDIR)/tfs/tlog.c \
+	$(SRCDIR)/fs/fs.c \
+	$(SRCDIR)/fs/tfs.c \
+	$(SRCDIR)/fs/tlog.c \
 	$(SRCDIR)/unix_process/unix_process_runtime.c
 LIBS-tfs-fuse=	-lfuse -lpthread
 
@@ -25,8 +27,9 @@ SRCS-mkfs= \
 	$(SRCDIR)/kernel/pagecache.c \
 	$(CURDIR)/mkfs.c \
 	$(RUNTIME) \
-	$(SRCDIR)/tfs/tfs.c \
-	$(SRCDIR)/tfs/tlog.c \
+	$(SRCDIR)/fs/fs.c \
+	$(SRCDIR)/fs/tfs.c \
+	$(SRCDIR)/fs/tlog.c \
 	$(SRCDIR)/unix_process/unix_process_runtime.c
 
 SRCS-vdsogen=	$(CURDIR)/vdsogen.c
@@ -36,7 +39,7 @@ CFLAGS+=-O3 \
 	-I$(SRCDIR) \
 	-I$(SRCDIR)/kernel \
 	-I$(SRCDIR)/runtime \
-	-I$(SRCDIR)/tfs \
+	-I$(SRCDIR)/fs \
 	-I$(SRCDIR)/unix_process
 
 all: $(PROGRAMS)
